chore: Update Docker configuration files
This commit updates the Docker configuration files by adding the `.env` file to the `.gitignore` and `.dockerignore` files. It also modifies the `docker-compose.yml` file to use port `8997` instead of `3000` for the Next.js application. Additionally, the `Dockerfile` is updated to use the specific version `20.15.0-alpine` of the Node.js image and to run the `npm run build` command instead of `npm run dev` for production.

```